pyproject.toml
[project]
name = "finetune-diffing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "backoff>=2.2.1",
    "fire>=0.7.0",
    "ipykernel>=6.29.5",
    "ipython>=9.0.2",
    "ipywidgets>=8.1.5",
    "matplotlib>=3.10.1",
    "mdmm>=0.1.3",
    "pip>=25.0.1",
    "python-dotenv>=1.1.0",
    "seaborn>=0.13.2",
    "tabulate>=0.9.0",
    "torch>=2.6.0",
    "transformer-heads>=0.2.2",
    "transformer-lens>=2.15.0",
    "transformers>=4.51.0",
    "unsloth>=2025.3.19",
    "vllm>=0.8.2",
]
